requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langgraph
langchain-core
python-dotenv
openai
pypdf
python-docx